Refactore some code of Wallet Screen
