+++ src/Popup/Router/Screens/Application/WalletScreen/CoinDashboard/SendScreen/FooterComponent.tsx
