+++ src/Popup/Router/Screens/Application/WalletScreen/CoinDashboard/SendScreen/SendDataFooterRow.tsx
