--- src/Popup/Router/Screens/Application/WalletScreen/CoinDashboard/SendScreenComponents/index.ts
@@ -1,2 +1,0 @@
-export {FooterComponent} from './FooterComponent';
-export {SendDataFooterRow, IFooterRowProps} from './SendDataFooterRow';
